# Conflicts:
#	src/main/java/com/project/chamjimayo/controller/dto/ApiStandardResponse.java
#	src/main/java/com/project/chamjimayo/controller/dto/ErrorCode.java
#	src/main/java/com/project/chamjimayo/domain/entity/Restroom.java
#	src/main/resources/config

+++ src/main/java/com/project/chamjimayo/domain/entity/BoardImage.java
@@ -1,6 +1,14 @@
 package com.project.chamjimayo.domain.entity;
 
-import javax.persistence.*;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.ManyToOne;
+import javax.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.ToString;
@@ -15,7 +23,7 @@ public class BoardImage extends BaseEntity {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "image_id")
-	private Integer imageId;
+	private Long imageId;
 
 	// 게시글 아이디 (어떤 게시글의 사진인가)
 	@ManyToOne(fetch = FetchType.LAZY)
@@ -25,5 +33,4 @@ public class BoardImage extends BaseEntity {
 	// 사진 url
 	@Column(name = "url")
 	private String url;
-
 }
